mito-ai: move websocket types to models
